move fork implementations to `fork` dir
