+++ crates/consensus/src/lib.rs
@@ -10,13 +10,10 @@ mod internal_prelude {
 }
 
 pub mod beacon;
-pub mod bellatrix;
 pub mod bls;
-pub mod capella;
 pub mod compute;
 pub mod config;
 pub mod context;
-pub mod deneb;
 pub mod errors;
 pub mod execution;
 pub mod fork;
